Nucleotide_Essentials.jl

Nucleotide_Essentials.jl

Data Types

Functions

Index

Data Types

Nucleotide_Essentials.FastqRecord — Type

Nucleotide_Essentials.FastqRecord

Components

ID: The unique sequence identifier associated with that entry

sequence: The nucleotide sequence of that entry

quality: The quality scores of that entry

filename: The original file name

Nucleotide_Essentials.FastaRecord — Type

Nucleotide_Essentials.FastaRecord

Components

ID: The unique sequence identifier associated with that entry

sequence: The nucleotide sequence of that entry

filename: The original file name

Functions

Nucleotide_Essentials.readFastq — Function

Nucleotide_Essentials.readFastq

readFastq(Path::String)

.fastq file => readFastq(Path) => FastqRecord(ID, sequence, quality, filename)

supported keyword arguments include:

'Path::String': The full or relative path to a .fastq file

Example:

# Supply the path to a .Fastq file that you would like to import
myfastq = readFastq("myfastq.fastq")

Nucleotide_Essentials.readFasta — Function

Nucleotide_Essentials.readFasta

readFasta(Path::String)

.fasta file => readFasta(Path) => FastaRecord(ID, sequence, filename)

supported keyword arguments include:

'Path::String': The full or relative path to a .fastq file

Example:

# Supply the path to a .Fastq file that you would like to import - it is recommended to include `;` in your command to prevent printing potentially large .fasta files in the REPL
myfasta = readFasta("myfasta.fasta");

Nucleotide_Essentials.PlotQuality — Function

Nucleotide_Essentials.PlotQuality

Returns a plot of the quality profile of a FastqRecord

This function plots a visual summary of the distribution of quality scores as a function of sequence position for the input fastq file(s).

The plotted lines show summary statistics at each sequence position:

green is the mean

dashed red lines are the 25th and 75th quantiles

supported keyword arguments include:

'Input::FastqRecord': The name of a FastqRecord for plotting

'QualityType::String' (optional): The system to use for parsing the fastq quality information. Potential quality encodings include:

Phred64 (defualt)

Ascii (not yet supported)

Phred33 (not yet supported)

'verbose::Bool' (optional): Whether or not to show some intermediary feedback on the progress of the function (default = false)

Example:

# A quality profile can be created by directly calling an already read FastqRecord
PlotQuality(myfastq)

# Alternatively, a quality profile can be created by nesting readFastq() inside of PlotQuality() 
PlotQuality(readFastq("myfastq.fastq"))

Nucleotide_Essentials.potential_mismatches — Function

Nucleotide_Essentials.potential_mismatches

Returns an Vector{Any} of potential barcodes with a single nucleotide change, including both deletions and substitutions

supported keyword arguments include:

'Path::String': The full or relative path to a .fastq file

'mismatch::Int64': The number of altered nucleotides to include (1 is only supported at this time)

Example:

potential_mismatches("GCGT", 1)
17-element Vector{Any}:
"GCGT"
"CCGT" 
"ACGT" 
"TCGT" 
"GGGT" 
"GAGT" 
"GTGT" 
"GCCT" 
"GCAT" 
"GCTT" 
"GCGG" 
"GCGC" 
"GCGA" 
"CGT"
"GGT"
"GCT" 
"GCG"

Nucleotide_Essentials.reverse_complement — Function

Nucleotide_Essentials.reverse_complement

Takes a string of nucleotide bases and returns the reverse complement of that string. Accepts inputs of String and SubString{String} (input from a FastqRecord)

supported keyword arguments include:

'sequence::Union{String, SubString{String}}': A string sequence of nucleotide bases or sequence entry from a FastqRecord

Example:

reverse_complement("ATCGT")
"ACGAT"

Nucleotide_Essentials.demultiplex_se — Function

Nucleotide_Essentials.demultiplex_se

Compares a list of provided barcodes with the provided multiplexed reads and separates the reads into individual .fastq files. If a barcode is found within the read, the barcode is removed from the sequence. The quality data of the reads is preserved and written to the outputted .fastq file. If a barcode is not found, the sequnce and quality is written to the unassigned .fastq file unchanged.

The mapping file must be either a .csv or .txt file with two columns. The first column heading must be SampleID and the second column heading must be BarcodeSequence.

EXAMPLE MAPPING FILE:

SampleID	BarcodeSequence

Sample1	Barcode1

Sample2	Barcode2

Sample3	Barcode3

Sample4	Barcode4

Sample5	Barcode5

Sample6	Barcode6

Sample7	Barcode7

Sample8	Barcode8

supported keyword arguments include:

'R1::String': Path to multiplexed reads

'Map::String': Path to the mapping file

'mismatch::Int64=0' (optional): Number of allowed mismatches in barcode. Potential options include 0 or 1. If 1 mismatch, computation time will significantly increase. Default is to allow for 0 mismatches (exact matches only).

'debug::Bool=false' (optional): If true, a log file will be created and debugging data will be printed while the function is running (default is false).

Example:

demultiplex_se("multiplexreads.fastq", "mapping_file.fastq")

Nucleotide_Essentials.demultiplex_pe — Function

Nucleotide_Essentials.demultiplex_pe

Compares a list of provided barcodes with the provided paired-end multiplexed reads and separates the reads into individual .fastq files. If a barcode is found within R1 reads, the barcode is removed from the sequence. The quality data of the reads is preserved and written to the outputted .fastq file. If a barcode is not found, the sequnce and quality is written to the R1 unassigned .fastq file unchanged. If a barcode is found within R2 reads, the barcode is removed from the sequence. The quality data of the reads is preserved and written to the outputted .fastq file. If a barcode is not found, the sequnce and quality is written to the R2 unassigned .fastq file unchanged.

Dual-indexed reads are not yet supported

The mapping file must be either a .csv or .txt file with two columns. The first column heading must be SampleID and the second column heading must be BarcodeSequence.

EXAMPLE MAPPING FILE:

SampleID	BarcodeSequence

Sample1	Barcode1

Sample2	Barcode2

Sample3	Barcode3

Sample4	Barcode4

Sample5	Barcode5

Sample6	Barcode6

Sample7	Barcode7

Sample8	Barcode8

supported keyword arguments include:

'R1::String': Path to forward multiplexed reads

'R2::String': Path to reverse multiplexed reads

'Map::String': Path to the mapping file

'mismatch::Int64=0' (optional): Number of allowed mismatches in barcode. Potential options include 0 or 1. If 1 mismatch, computation time will significantly increase. Default is to allow for 0 mismatches (exact matches only).

'debug::Bool=false' (optional): If true, a log file will be created and debugging data will be printed while the function is running (default is false).

Example:

demultiplex_pe("forward_multiplexreads.fastq", "reverse_multiplexreads.fastq", "mapping_file.fastq")

Index

Nucleotide_Essentials.FastaRecord

Nucleotide_Essentials.FastqRecord

Nucleotide_Essentials.PlotQuality

Nucleotide_Essentials.demultiplex_pe

Nucleotide_Essentials.demultiplex_se

Nucleotide_Essentials.potential_mismatches

Nucleotide_Essentials.readFasta

Nucleotide_Essentials.readFastq

Nucleotide_Essentials.reverse_complement